disable fallbacks for old compilers in the Docker build
Normally the `blake3` crate will silently skip building AVX-512 code if
it detects that the C compiler doesn't support that. In the manylinux
Docker image we use, that should never happen. But just in case, set
BLAKE3_CI=1 to make sure that if it ever does happen, we fail the build
instead of silently publishing an artifact with that's missing its
AVX-512 implementation.