add a Mutex to the Rust implementation
This matches the previous commit for the C implementation and the
standard practice in hashlib. Because we release the GIL during hashing,
it's possible that another thread could race to update the same hasher.
Putting the underlying hasher inside a Mutex prevents this race.

Note that the result is not entirely race-free. It's still possible for
another thread to update the *buffer* while this code is treating it as
a &[u8]. The effect of this in practice is "probably" just junk bytes,
but strictly speaking it's still unsound.