switch from tp_init to tp_new
tp_init (equivalent to __init__) is allowed to be called more than once,
or even not at all. Now that it's responsible for initializing
self->lock, that would lead to a memory leak and/or memory corruption,
and we need to define tp_new instead.

This change fixes a bug where we were incorrectly freeing self->lock
even if tp_init returned early without initializing it. This was causing
segfaults in CI on macOS.