compile c_impl to a universal binary on macOS
I think setuptools is copying the flags that Python itself was built
with, and as of Python 3.11 that means it's trying to build a universal
binary. This goes poorly, because it doesn't know how to build different
files for x86 and for ARM, so we need to do it explicitly ourselves.