Fix casts in C implementation for GCC 14
Fix casts performed in the C extension module to make it build with
GCC 14.  This involves adding two missing casts from `Blake3Object *`
to `PyObject *` for return values, and fixing casts in method
definitions to use `PyCFunction` (i.e. the type expected by the struct)
rather than `PyCFunctionWithKeywords` (i.e. the type that's actually
used by the function, and that wouldn't require a cast in the first
place).

Fixes #43