set gil_used = false
This gets rid of the following warning on free-threaded builds:

```
<frozen importlib._bootstrap>:491: RuntimeWarning: The global
interpreter lock (GIL) has been enabled to load module 'blake3.blake3',
which has not declared that it can run safely without the GIL. To
override this behavior and keep the GIL disabled (at your own risk), run
with PYTHON_GIL=0 or -Xgil=0.
```

As part of this, refactor `Blake3Class` to be a "frozen" class. It
already managed its `blake3::Hasher` with a `Mutex`, this doesn't change
much internally, mostly just method signatures. The PyO3 docs encourage
this approach, and it avoids raising inconsistent and potentially
confusing "already borrowed" exceptions in programs with race conditions
(even though such a program is probably still buggy).